Dont try to retrieve the file if its already there, and remove inline import

+++ pavement.py
@@ -19,6 +19,7 @@ from shutil import move
 import zipfile
 import tarfile
 import urllib
+from urllib import urlretrieve
 import glob
 
 assert sys.version_info >= (2,6), \
@@ -152,8 +153,8 @@ def post_bootstrap(options):
 #TODO Move svn urls out to a config file
 
 def grab(src, dest):
-    from urllib import urlretrieve
-    urlretrieve(str(src), str(dest))
+    if not os.path.exists(dest):
+        urlretrieve(str(src), str(dest))
 
 @task
 def setup_gs_data(options):
@@ -315,7 +316,6 @@ def create_version_name():
 @task
 def make_devkit(options):
     import virtualenv
-    from urllib import urlretrieve
 
     (path("package") / "devkit" / "share").makedirs()
     pip_bundle("package/devkit/share/geonode-core.pybundle -r shared/devkit.requirements")
